requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.9
dj-database-url==3.0.0
Django==5.2.3
filelock==3.18.0
Flask==3.1.0
fonttools==4.58.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mysql-connector==2.2.9
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
whitenoise==6.9.0
